requirements.txt
colorama==0.4.6
decorator==5.2.1
ffmpeg-python==0.2.0
future==1.0.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.1.2
numpy==2.2.4
pillow==10.4.0
proglog==0.1.11
pydub==0.25.1
python-dotenv==1.1.0
tqdm==4.67.1
